Added "set model" debug command (requires manual respawn for now)

+++ src/world/Actor/Player.h
@@ -816,7 +816,7 @@ namespace Sapphire::Entity
 	
 	uint16_t getModelChara() const;
 	
-	void setModelChara( uint8_t modelChara );
+	void setModelChara( uint16_t modelChara );
 
     // Content Finder handling
     //////////////////////////////////////////////////////////////////////////////////////////////////////
